bugfix major: handle apply on cons cells
we tried to make a shortcut: when applying a list, just pick out the
instruction and put the cdr of the list back on the exec stack. this
meant we didn't have to do a complete "append".

this has a big problem, of course. consider ((#f exec.if 1 0)):

- the list would mean that there's only 1 element on the stack
- the exec.if pop-check ensures that we need two elements

this patch fixes this. apply on a pair will push the entire list onto
the exec stack. it still does does an optimization that may/may not
work: it allows improper list and will execute non-'() tails too.

this means we don't need "ensure-list" from schush.ss but maybe this
will cause trouble too.