test cleanup: use schush.ss
use original schush.ss from Lee Spector as benchmark comparison. my port
to Chicken isn't really interesting (it's also a lot slower, for some
reason ...)

this means we don't get a speedup printout, but we can look at the
timings of each. I'm getting 2400ms for mzscheme and 145ms for the
C-generated VM. it's not that big of an improvement, actually. 10x
faster isn't worth the trouble of losing the flexibility of the Lisp
languages. and the C runtime will grow too. we need a GC and things like
that so this is an ambition project.